General code improvement and updated bloc.
Once again, I would like to address the issue of if else. During production, excessive usage of it becomes a nightmare.
It is fine to use it but not for every state you handle. I have editted changed bloc state class so that is no longer necessary to use if else.
Whenever you dont need to use context, you can just put a place holder _. this way if your code is long, you will tell the reader that you wont use context in that scope.